requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.5.2
    # via django
backports-zoneinfo==0.2.1
    # via django
boto3==1.26.4
    # via django-s3-folder-storage
botocore==1.29.4
    # via
    #   boto3
    #   s3transfer
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==2.1.1
    # via requests
collectfast==2.2.0
    # via -r requirements.in
dj-database-url==1.0.0
    # via -r requirements.in
dj-rest-auth==2.2.5
    # via -r requirements.in
django==4.1.4
    # via
    #   -r requirements.in
    #   collectfast
    #   dj-database-url
    #   dj-rest-auth
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-phonenumber-field
    #   django-storages
    #   djangorestframework
    #   djangorestframework-simplejwt
django-cleanup==6.0.0
    # via -r requirements.in
django-cors-headers==3.13.0
    # via -r requirements.in
django-debug-toolbar==3.7.0
    # via -r requirements.in
django-extensions==3.2.1
    # via -r requirements.in
django-phonenumber-field[phonenumberslite]==7.0.1
    # via -r requirements.in
django-s3-folder-storage==0.5
    # via -r requirements.in
django-storages==1.13.1
    # via
    #   collectfast
    #   django-s3-folder-storage
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   dj-rest-auth
    #   djangorestframework-simplejwt
djangorestframework-camel-case==1.3.0
    # via -r requirements.in
djangorestframework-simplejwt==5.2.2
    # via -r requirements.in
freezegun==1.2.2
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
idna==3.4
    # via requests
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
phonenumberslite==8.13.3
    # via django-phonenumber-field
psycopg2-binary==2.9.5
    # via -r requirements.in
pyjwt==2.6.0
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
python-dateutil==2.8.2
    # via
    #   botocore
    #   freezegun
python-decouple==3.6
    # via -r requirements.in
pytz==2022.6
    # via djangorestframework
requests==2.28.1
    # via -r requirements.in
s3transfer==0.6.0
    # via boto3
sentry-sdk==1.10.1
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.3
    # via
    #   django
    #   django-debug-toolbar
typing-extensions==4.4.0
    # via collectfast
urllib3==1.26.12
    # via
    #   botocore
    #   requests
    #   sentry-sdk
validate-docbr==1.10.0
    # via -r requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
